Refactor ai utils (#297)
