--- vizro-ai/src/vizro_ai/utils/__init__.py
@@ -1,3 +1,0 @@
-from .safeguard import _safeguard_check
-
-__all__ = ["_safeguard_check"]
